pyfoo: Implement PyFormatter
This is a special in-process Formatter that
gives you a pointer to a python dict instead
of serializing the data.  It allows you to
pass dumpable C++ objects up into embedded
python code comparatively efficiently.

Ultimately one may still want explicit Python
wrappers for some classes, but this avoids the
need to write them everywhere: the python code
can just behave as if it was passed some JSON, but
we've skipped the intermediate serialization step.

Signed-off-by: John Spray <john.spray@redhat.com>